rename dsutil to typeutil
